+++ RobWork/test/geometry/GeometryUtilTest.cpp
@@ -63,7 +63,7 @@ BOOST_AUTO_TEST_CASE( EstimateVolumeTest ){
 		// Cylinder test
 		const double radius = 0.02;
 		const double height = 0.1;
-		const Cylinder cyl(radius, height);
+		const Cylinder cyl(static_cast<float>(radius), static_cast<float>(height));
 		const double volEst = GeometryUtil::estimateVolume(*cyl.createMesh(26));
 		BOOST_CHECK_CLOSE(volEst,Pi*radius*radius*height,1);
 	}
@@ -90,7 +90,7 @@ BOOST_AUTO_TEST_CASE( EstimateVolumeTest ){
 		const double radius = 0.02;
 		const double thickness = 0.005;
 		const double height = 0.1;
-		const Tube tube(radius, thickness, height);
+		const Tube tube(static_cast<float>(radius), static_cast<float>(thickness), static_cast<float>(height));
 		const double volEst = GeometryUtil::estimateVolume(*tube.createMesh(26));
 		BOOST_CHECK_CLOSE(volEst,Pi*height*((radius+thickness)*(radius+thickness)-radius*radius),1);
 	}
@@ -177,7 +177,7 @@ BOOST_AUTO_TEST_CASE( EstimateCOGTest ){
 		// Cylinder test
 		const double radius = 0.02;
 		const double height = 0.1;
-		const Cylinder cyl(radius, height);
+		const Cylinder cyl(static_cast<float>(radius), static_cast<float>(height));
 		const Vector3D<> cogEst = GeometryUtil::estimateCOG(*cyl.createMesh(26));
 		BOOST_CHECK_SMALL(dot(cogEst,Vector3D<>::x()),std::numeric_limits<double>::epsilon());
 		BOOST_CHECK_SMALL(dot(cogEst,Vector3D<>::y()),std::numeric_limits<double>::epsilon());
@@ -210,7 +210,7 @@ BOOST_AUTO_TEST_CASE( EstimateCOGTest ){
 		const double radius = 0.02;
 		const double thickness = 0.005;
 		const double height = 0.1;
-		const Tube tube(radius, thickness, height);
+		const Tube tube(static_cast<float>(radius), static_cast<float>(thickness), static_cast<float>(height));
 		const Vector3D<> cogEst = GeometryUtil::estimateCOG(*tube.createMesh(26));
 		BOOST_CHECK_SMALL(dot(cogEst,Vector3D<>::x()),std::numeric_limits<double>::epsilon());
 		BOOST_CHECK_SMALL(dot(cogEst,Vector3D<>::y()),std::numeric_limits<double>::epsilon());
@@ -316,7 +316,7 @@ BOOST_AUTO_TEST_CASE( EstimateInertiaTest ){
 		// Cylinder test
 		const double radius = 0.02;
 		const double height = 0.1;
-		const Cylinder cyl(radius, height);
+		const Cylinder cyl(static_cast<float>(radius), static_cast<float>(height));
 		const std::vector<Geometry::Ptr> geoms(1,ownedPtr(new Geometry(cyl.createMesh(26))));
 		const InertiaMatrix<> inertiaEst = GeometryUtil::estimateInertia(mass, geoms, ref);
 		const double Ix = mass*(3*radius*radius+height*height)/12.;
@@ -361,7 +361,7 @@ BOOST_AUTO_TEST_CASE( EstimateInertiaTest ){
 		const double radius = 0.02;
 		const double thickness = 0.005;
 		const double height = 0.1;
-		const Tube tube(radius, thickness, height);
+		const Tube tube(static_cast<float>(radius), static_cast<float>(thickness), static_cast<float>(height));
 		const std::vector<Geometry::Ptr> geoms(1,ownedPtr(new Geometry(tube.createMesh(26))));
 		const InertiaMatrix<> inertiaEst = GeometryUtil::estimateInertia(mass, geoms, ref);
 		const double Ix = mass*(3*radius*radius+3*(radius+thickness)*(radius+thickness)+height*height)/12.;
